+++ packages/editor/js/extensions/libs/size/styles.js
@@ -17,14 +17,14 @@ import {
 	getCompatibleBlockCssSelector,
 	computedCssDeclarations,
 } from '../../../style-engine';
-import { getBlockSupportCategory, getBlockSupportFallback } from '../../utils';
-
-const supports = getBlockSupportCategory('size');
+import { getBlockSupportFallback } from '../../utils';
+import { WidthGenerator } from './css-generators/width-generator';
 
 export const SizeStyles = ({
 	state,
 	config,
 	clientId,
+	supports,
 	blockName,
 	masterState,
 	currentBlock,
@@ -51,7 +51,9 @@ export const SizeStyles = ({
 		clientId,
 		blockName,
 		attributes: currentBlockAttributes,
+		supports,
 	};
+
 	const sharedParams = {
 		...props,
 		state,
@@ -72,20 +74,6 @@ export const SizeStyles = ({
 	const styleGroup: Array<CssRule> = [];
 
 	if (isActiveField(blockeraWidth) && currentBlockAttributes?.blockeraWidth) {
-		const width = getValueAddonRealValue(
-			currentBlockAttributes.blockeraWidth
-		);
-		let value = '';
-
-		if (width !== attributes.blockeraWidth.default) {
-			value = width;
-		} else if (
-			!isUndefined(currentBlockAttributes.width) &&
-			!isEmpty(currentBlockAttributes.width)
-		) {
-			value = currentBlockAttributes.width;
-		}
-
 		const pickedSelector = getCompatibleBlockCssSelector({
 			...sharedParams,
 			query: 'blockeraWidth',
@@ -102,10 +90,8 @@ export const SizeStyles = ({
 				{
 					blockeraWidth: [
 						{
-							...staticDefinitionParams,
-							properties: {
-								width: value,
-							},
+							type: 'function',
+							function: WidthGenerator,
 						},
 					],
 				},