+++ packages/editor/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Various bug fixes.
 - Fixed an issue where the style engine generated incorrect selectors on both server and client sides.
 - Block section settings are now hidden if there are no taggable features in the section.
+- The Width feature should use `flex-basis` in the Column Block. 
 - Fixed an issue where the toggle action in the block section was not working correctly.
 - Fixed an issue where the block section focus, collapse and expand actions were not being updated correctly as a unified state around the editor for all blocks.
 - Fixed the Flex Direction not working properly sometimes.